OK so I added my pseudo code. I set it as the controller running as much as possible, I did not set it up with a message system. the controller just calls the function through the desired object.
Update.cpp is kind of the main game loop.

Not sure if I went in depth enough.

I also tried to keep a tally of input and (UI) and screen objects (HUD) by making notes for these "classes"

most of the functions I was able to have something in them to explain what they might do so look at definition on functions I hope it this thing is user friendly enough.